# Conflicts:
#	src/app/Main.java
#	src/data_access/YouTubeDataAccess.java
#	src/entities/Video.java
#	src/entities/VideoFactory.java
#	src/views/HomeView.java

+++ src/entities/VideoFactory.java
@@ -9,7 +9,7 @@ public class VideoFactory {
 
     Video create(String videoId, String channelName, String title,
                  String description, DateTime videoPublishDate,
-                 BigInteger viewCount, BigInteger likeCount, BigInteger commentCount) {
+                 int viewCount, int likeCount, int commentCount) {
         return new Video(videoId, channelName, title,
                 description, videoPublishDate,
                 viewCount, likeCount, commentCount);
